rename timer -> stopwatch everywhere
commit-id:9f1a2986
